@@ -2,4 +2,4 @@ module.exports = {
     presets: [
         '@vue/cli-plugin-babel/preset'
     ]
-};
+}
