@@ -1,18 +1,18 @@
 <template>
     <div :class="{unread : chat.isUnread}" class="last-msg flex-grow-1 d-flex align-items-center">
-        <template v-if="chat.lastMsg">
-            <MessageIconStatus :ack="chat.lastMsg.ack" class="icon-status" v-if="chat.lastMsg.id.fromMe"/>
+        <template v-if="lastMsg">
+            <MessageIconStatus :ack="lastMsg.ack" class="icon-status" v-if="lastMsg.id.fromMe"/>
 
-            <span :inner-html.prop="`${formattedNameLastMsg}: ` | emojify"
+            <span v-html="`${formattedNameLastMsg}: `"
                   v-if="chat.isGroup && formattedNameLastMsg"></span>
 
-            <MessageBody :lastMsg="chat.lastMsg"/>
+            <MessageBody :lastMsg="lastMsg"/>
         </template>
     </div>
 </template>
 
 <script>
-
+import filters from '@/filters';
 import MessageBody from './messageBody/MessageBody.vue';
 import MessageIconStatus from '@/components/shared/messageIconStatus/MessageIconStatus.vue';
 
@@ -22,30 +22,34 @@ export default {
         MessageIconStatus,
         MessageBody
     },
-    asyncComputed: {
-        formattedNameLastMsg: {
-            async get () {
-                let senderObj = await this.chat.lastMsg.senderObj();
-                return senderObj.formattedName;
-            },
-            default () {
-                return false;
-            }
-        }
-    },
     props: {
         chat: {
             type: Object,
             required: true
         }
     },
+    computed: {
+        lastMsg () {
+            return this.chat.lastMsg;
+        }
+    },
     data () {
-        return {};
+        return {
+            formattedNameLastMsg: null
+        };
     },
     created () {
-
+        this.loadFormattedName();
     },
-    watch: {}
+    watch: {
+        'lastMsg': 'loadFormattedName'
+    },
+    methods: {
+        async loadFormattedName () {
+            let senderObj = await this.chat.lastMsg.senderObj();
+            this.formattedNameLastMsg = filters.emojify(senderObj.formattedName);
+        }
+    }
 };
 </script>
 
