@@ -32,16 +32,18 @@
                 <!--<b-dropdown-item to="/changenumber" v-if="user.canCreateOperator">Alterar Número</b-dropdown-item>-->
                 <b-dropdown-item to="/changepassword">Alterar Senha</b-dropdown-item>
                 <b-dropdown-item to="/newoperator" v-if="user.canCreateOperator">Novo Operador</b-dropdown-item>
-                <b-dropdown-item to="/manageoperators" v-if="user.canCreateOperator">Gerenciar Operadores</b-dropdown-item>
+                <b-dropdown-item to="/manageoperators" v-if="user.canCreateOperator">Gerenciar Operadores
+                </b-dropdown-item>
                 <b-dropdown-form class="text-nowrap" v-if="user.canCreateOperator">
-                    <b-form-checkbox name="check-button" switch v-model="user.configuracao.enviarNomeOperadores">
+                    <!--<b-form-checkbox name="check-button" switch v-model="user.configuracao.enviarNomeOperadores">
                         Enviar Nome Operadores
-                    </b-form-checkbox>
+                    </b-form-checkbox>-->
                 </b-dropdown-form>
                 <b-dropdown-form class="text-nowrap" v-if="user.canCreateOperator">
+                    <!--TODO checkBox
                     <b-form-checkbox name="check-button" switch v-model="user.configuracao.operadorPodeExcluirMsg">
-                        Operadores podem excluir mensagens?
-                    </b-form-checkbox>
+                       Operadores podem excluir mensagens?
+                   </b-form-checkbox>-->
                 </b-dropdown-form>
             </b-dropdown>
         </div>
@@ -49,7 +51,7 @@
 </template>
 
 <script>
-import { mapState, mapMutations } from 'vuex';
+import { mapMutations, mapState } from 'vuex';
 import api from '@/api';
 
 export default {
