@@ -7,7 +7,8 @@
                     <div class="select-msg-check"></div>
                 </div>
             </div>
-            <div :class="showTail ? 'tail' : ''" class="message-body" v-b-hover="handleHover">
+            <!--TODO v-b-hover="handleHover" -->
+            <div :class="showTail ? 'tail' : ''" class="message-body">
                 <div class="tail-container" v-if="showTail"></div>
                 <div
                     class="msg-menu"
@@ -42,12 +43,12 @@
                      v-if="activeChat.isGroup && !msg.id.fromMe && showTail">
                     <div :style="{color: colorMsgGroup}" @click="handleClick" class="btn-link"
                          v-if="name">
-                        {{name | emojify}}
+                        {{ filters.emojify(name) }}
                     </div>
 
                     <div class="d-flex justify-content-between" v-else>
                     <span :style="{color: colorMsgGroup}" @click="handleClick"
-                          class="btn-link number">{{formattedName | emojify}}</span>
+                          class="btn-link number">{{ filters.emojify(formattedName) }}</span>
                         <span class="name">~{{pushName}}</span>
                     </div>
                 </div>
@@ -73,6 +74,8 @@
 </template>
 
 <script>
+import filters from '@/filters';
+import { asyncComputed } from '@/AsyncComputed.ts';
 import { mapActions, mapMutations, mapState } from 'vuex';
 import MessageText from '@/components/shared/messageText/MessageText.vue';
 import MessagePhoto from '@/components/shared/messagePhoto/MessagePhoto.vue';
@@ -106,26 +109,29 @@ export default {
     },
     data () {
         return {
+            filters,
             showMenuIcon: false,
             menuAberto: false
         };
     },
     mounted () {
-        this.$root.$on('keyDown', (evt) => {
+        //TODO remove selected msg on esc press
+        /*this.$root.$on('keyDown', (evt) => {
             if (this.showSelectMsgs) {
                 evt.preventDefault();
                 if (evt.key === 'Escape') {
                     this.SET_SELECT_MSGS({ show: false });
                 }
             }
-        });
+        });*/
     },
     watch: {
         'activeChat.quotedMsg': function (val) {
             if (val === this.msg) {
-                this.msg.blink = true;
+                //TODO: blink msg
+                //this.msg.blink = true;
                 setTimeout(() => {
-                    this.msg.blink = false;
+                    //this.msg.blink = false;
                 }, 1500);
             }
         },
@@ -162,56 +168,43 @@ export default {
             return this.selectMsgs.msgs.includes(this.msg);
         }
     },
-    asyncComputed: {
-        colorMsgGroup: {
-            async get () {
-                if (this.activeChat.isGroup) {
-                    let senderObj = await this.msg.senderObj();
-                    return this.activeChat.getColor(senderObj.id);
-                }
-                return false;
-            },
-            default () {
-                return randomColor().hexString();
-            }
-        },
-        name: {
-            async get () {
+    setup () {
+        const colorMsgGroup = asyncComputed(async () => {
+            if (this.activeChat.isGroup) {
                 let senderObj = await this.msg.senderObj();
-                return senderObj.name;
-            },
-            default () {
-                return '';
+                return this.activeChat.getColor(senderObj.id);
             }
-        },
-        pushName: {
-            async get () {
-                let senderObj = await this.msg.senderObj();
-                return senderObj.pushname;
-            },
-            default () {
-                return '';
-            }
-        },
-        formattedName: {
-            async get () {
-                let senderObj = await this.msg.senderObj();
-                return senderObj.formattedName;
-            },
-            default () {
-                return '';
-            }
-        },
-        showTail: {
-            async get () {
-                let senderObj = await this.msg.senderObj();
-                let previousMsgSenderObj = this.previousMsg ? await this.previousMsg.senderObj() : false;
-                return !this.previousMsg || !senderObj || !previousMsgSenderObj || (this.previousMsg.isSticker && !this.msg.isSticker) || senderObj.id !== previousMsgSenderObj.id || this.msg.hasQuotedMsg || this.msg.isSticker || this.msg.fomattedDate !== this.previousMsg.fomattedDate;
-            },
-            default () {
-                return false;
-            }
-        }
+            return false;
+        }, { default: randomColor().hexString(), lazy: true });
+
+        const name = asyncComputed(async () => {
+            let senderObj = await this.msg.senderObj();
+            return senderObj.name;
+        }, { default: '', lazy: true });
+
+        const pushName = asyncComputed(async () => {
+            let senderObj = await this.msg.senderObj();
+            return senderObj.pushname;
+        }, { default: '', lazy: true });
+
+        const formattedName = asyncComputed(async () => {
+            let senderObj = await this.msg.senderObj();
+            return senderObj.formattedName;
+        }, { default: '', lazy: true });
+
+        const showTail = asyncComputed(async () => {
+            let senderObj = await this.msg.senderObj();
+            let previousMsgSenderObj = this.previousMsg ? await this.previousMsg.senderObj() : false;
+            return !this.previousMsg || !senderObj || !previousMsgSenderObj || (this.previousMsg.isSticker && !this.msg.isSticker) || senderObj.id !== previousMsgSenderObj.id || this.msg.hasQuotedMsg || this.msg.isSticker || this.msg.fomattedDate !== this.previousMsg.fomattedDate;
+        }, { default: false, lazy: true });
+
+        return {
+            colorMsgGroup,
+            name,
+            pushName,
+            formattedName,
+            showTail,
+        };
     },
     methods: {
         ...mapMutations(['SET_ACTIVE_CHAT', 'SET_SELECT_MSGS', 'TOGGLE_SELECT_MSG']),
@@ -221,19 +214,19 @@ export default {
             this.showMenuIcon = isHover;
         },
 
-        handleShowMenu (evt) {
+        handleShowMenu () {
             this.menuAberto = true;
         },
 
-        handleHideMenu (evt) {
+        handleHideMenu () {
             this.menuAberto = false;
         },
 
-        handleClickAnswer (evt) {
+        handleClickAnswer () {
             this.activeChat.quotedMsg = this.msg;
         },
 
-        handleClickForward (evt) {
+        handleClickForward () {
             this.SET_SELECT_MSGS({ show: true });
         },
 
@@ -268,6 +261,7 @@ export default {
 
                 document.body.removeChild(element);
             }).catch(reason => {
+                console.error(reason);
                 alert('Falha ao baixar o arquivo, atualize a pagina e tente novamente.');
             });
         }
@@ -386,7 +380,7 @@ export default {
     right: -12px;
 }
 
-.message-out.blink .message-body, .message-out.blink >>> .time {
+.message-out.blink .message-body, .message-out.blink ::v-deep(.time) {
     animation-name: blink-out;
     animation-duration: 1500ms;
     animation-timing-function: ease-in-out;
@@ -394,7 +388,7 @@ export default {
     animation-play-state: running;
 }
 
-.message-in.blink .message-body, .message-in.blink .identify-msg-group, .message-in.blink >>> .time {
+.message-in.blink .message-body, .message-in.blink .identify-msg-group, .message-in.blink ::v-deep(.time) {
     animation-name: blink-in;
     animation-duration: 1500ms;
     animation-timing-function: ease-in-out;
