@@ -1,5 +1,9 @@
 <template>
-    <div class="message-gif" v-b-visible.once="onVisible">
+    <div class="message-gif" v-observe-visibility="{
+         throttle: 300,
+         callback: onVisible,
+         once: true
+    }">
         <div class="gif-container">
             <div class="box-preview blur" v-if="!srcVideo">
                 <img :src=" 'data:image/jpeg;base64,'+ msg.body" alt="body">
@@ -135,7 +139,7 @@ video {
     bottom: 3px;
 }
 
-.box-caption >>> .mention-symbol {
+.box-caption ::v-deep(.mention-symbol) {
     color: rgba(0, 0, 0, 0.25)
 }
 </style>
