@@ -108,7 +108,7 @@
 </template>
 
 <script>
-import { mapState, mapActions } from 'vuex';
+import { mapActions, mapState } from 'vuex';
 import LoadginSpinner from '../components/shared/loadingSpinner/LoadingSpinner';
 
 export default {
