@@ -1,40 +1,15 @@
-const webpack = require('webpack');
+const WorkerPlugin = require('worker-plugin');
 
 module.exports = {
     configureWebpack: {
+        output: {
+            globalObject: 'this'
+        },
         plugins: [
-            new webpack.ProvidePlugin({
-                mapboxgl: 'mapbox-gl'
-            })
+            new WorkerPlugin()
         ]
     },
     chainWebpack: config => {
-        config.module
-            .rule('vue')
-            .use('vue-loader')
-            .loader('vue-loader')
-            .tap(options => {
-                options['transformAssetUrls'] = {
-                    video: ['src', 'poster'],
-                    source: 'src',
-                    img: 'src',
-                    image: 'xlink:href',
-                    'b-img': 'src',
-                    'b-img-lazy': ['src', 'blank-src'],
-                    'b-card': 'img-src',
-                    'b-card-img': 'src',
-                    'b-card-img-lazy': ['src', 'blank-src'],
-                    'b-carousel-slide': 'img-src',
-                    'b-embed': 'src'
-                };
-                return options;
-            });
-        config.module
-            .rule('worker-loader')
-            .test(/\.worker\.js$/)
-            .use('worker-loader')
-            .loader('worker-loader')
-            .end();
         config.module
             .rule('file-loader')
             .test(/\.wasm$/)
